// frontend/src/types/Product.ts
export type Product = {
  Id: number;
  BrandId: number;
  BrandName:string;
  CategoryName:string;
  UrlProduct:string;
  CategoryId: number;
  Price:number;
  UserId: number;
  CartId: number;
  SpecificationId: number;
  Name: string;
  Slug: string;
  CretedDate: Date;
  UpdatedDate: Date;
  IsDeleted: boolean;
};